move layouts into scaffolder-react
