+++ plugins/scaffolder-react/src/next/types.ts
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 import type { FormProps as SchemaFormProps } from '@rjsf/core-v5';
-
 /**
  * Any `@rjsf/core` form properties that are publicly exposed to the `NextScaffolderpage`
  *
- * @alpha
+ * @public
  */
 export type FormProps = Pick<
   SchemaFormProps,
